+++ core-fns/github/featurerequest.go
@@ -1,0 +1,8 @@
+package github
+
+import "fmt"
+
+// requests the feature to the gievn url
+func FeatureRequest() (url string,err error) {
+	return fmt.Sprintf(`%s?template=feature_request.md`,GithubReportURL),nil
+}